requirements.txt
Flask
Gunicorn
BytesIO
base64
matplotlib.pyplot
numpy
requests
json
